Fix final de Visualizaciones. Creación de Makefile
