+++ EDA/🏠_EDA_home.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import streamlit as st
-from pages.utils import import_data
+from pages.utils import create_histogram, import_data, summary_data
 
 st.set_page_config(layout="wide")
 st.markdown(
@@ -22,43 +22,6 @@ if "EXCLUDE_COLUMNS" not in st.session_state:
     st.session_state.EXCLUDE_COLUMNS = ["not_null", "not_null_perc"]
 if "df_dict" not in st.session_state:
     st.session_state.df_dict = import_data()
-
-
-def create_histogram(event, color):
-    plot = (
-        st.session_state.df_dict[event]["not_null_perc"]
-        .plot(
-            kind="hist",
-            title="Histograma de % de Valores No Nulos",
-            color=color,
-            edgecolor="black",
-            bins=100,
-        )
-        .figure
-    )
-    return st.pyplot(plot, clear_figure=True)
-
-
-def summary_data(evento):
-    container = st.container()
-    container.header(evento)
-    container.metric(
-        label="Dimensiones", value=f"{st.session_state.df_dict[evento].shape}"
-    )
-    container.metric(
-        label="Mediciones Promedio",
-        value=f"{st.session_state.df_dict[evento]['not_null'].mean():.0f}/{st.session_state.df_dict[evento]['not_null_perc'].mean():.0f}%",
-    )
-    container.metric(
-        label="Mínimo de Mediciones",
-        value=f"{st.session_state.df_dict[evento]['not_null'].min():.0f}/{st.session_state.df_dict[evento]['not_null_perc'].min():.0f}%",
-    )
-    container.metric(
-        label="Máximo de Mediciones",
-        value=f"{st.session_state.df_dict[evento]['not_null'].max():.0f}/{st.session_state.df_dict[evento]['not_null_perc'].max():.0f}%",
-    )
-    return container
-
 
 col1, col2 = st.columns(2)
 df = pd.DataFrame(
