+++ Makefile
@@ -1,0 +1,4 @@
+.PHONY: dashboard
+
+dashboard:
+	streamlit run EDA/🏠_EDA_home.py
